fix(plugins): add missing ExtensionsPlugin struct and fix module structure
- Add missing 'ExtensionsPlugin' struct definition to 'extensions.rs'
- Move 'types.rs' to 'extensions/types.rs' (correct module path)
- Make 'types' module public ('pub mod types;')
- Re-export types ('pub use types::*;')
- Fix unused variable 'name' in 'clear' handler (implement correct logic)
- Remove unused imports ('Deserialize', 'Serialize')

Fixes compilation errors: E0422, E0425, E0433
